RF: Make old gui a legacy module
