--- psychopy/gui2/wxgui.py
@@ -1,6 +1,0 @@
-from psychopy.gui2.base import BaseDlg
-import wx
-
-
-class Dlg(wx.Dialog, BaseDlg):
-    pass
